blog/v1/serializers.py
from rest_framework import serializers
from ..models import Post, Comment

class CommentSerializerV1(serializers.ModelSerializer):
    author = serializers.StringRelatedField(read_only=True)

    class Meta:
        model = Comment
        fields = ['id', 'post', 'content', 'author', 'timestamp']

class PostSerializerV1(serializers.ModelSerializer):
    author = serializers.StringRelatedField(read_only=True)

    class Meta:
        model = Post
        fields = ['id', 'title', 'content', 'author', 'timestamp']
