requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
pillow==11.0.0
djangorestframework==3.15.2
drf-yasg==1.21.8